Merged /crujisim/branches/cleansim r1787-1908
This branch makes the simulation engine an independent object (GTA) to which clients connect. This was done by isolating the relevant code segments from Simulador.py, which is still present on the directory tree, but only for reference, since it is not invoked anymore.

Now the simulation may be controlled remotely using the pseudopilot interface. The pseudopilot interface my be invoked using RemotePP.py

A good cleanup has been made in the area of local maps. It's easier now to add points to the FIR files, and some of the maps are generated automatically from the known information (eg. the SID and STAR maps)

The label separation code is now run on it's own thread, and should not interrupt the simulation while running.

We are now positive that we have no more leaks, since we can see how the objects for the display and the simulation are deleted in between simulation runs.


Eliminated MainMenu.py, the launcher for